fix assess download test for type

+++ tests/testthat/test_assess_downloads.R
@@ -3,7 +3,7 @@ test_that("assess_downloads returns expected result for source package", {
     assess_source_good$downloads_1yr,
     c("pkg_metric_downloads_1yr", "pkg_metric", "integer"))
 
-  expect_type(assess_source_good$downloads_1yr, "integer")
+  expect_type(assess_source_good$downloads_1yr, "double")
   expect_lte(assess_source_good$downloads_1yr, 1)
   expect_gte(assess_source_good$downloads_1yr, 0)
 })
